+++ LSTM_model/data/data3.csv
@@ -1,0 +1,7 @@
+22:39:58.525596,57.5, 19.8, 19.8, 5000, 26
+22:40:00.526924,57.5, 19.8, 19.9, 5000, 26
+22:40:02.527584,57.5, 19.8, 19.9, 5000, 26
+22:40:24.543147,57.5, 19.8, 19.8, 5000, 26
+22:40:26.543416,57.5, 19.8, 19.8, 5000, 26
+22:40:28.547098,57.5, 19.8, 19.8, 5000, 26
+22:40:30.547046,57.4375, 19.8, 19.8, 5000, 26
