+++ sensordata_web/static/highcharts-more.js
