+++ sensordata_web/static/highcharts-more1.js
