+++ sensordata_web/static/highstock.js
