+++ sensordata_web/static/highstock1.js
