+++ sensordata_web/templates/MH_Z19T.html
@@ -1,0 +1,37 @@
+<!DOCTYPE HTML>
+<html>
+	<head>
+		<meta http-equiv="Content-Type" content="text/html; charset=utf-8">
+		<title>MH_Z19T</title>
+        <script src="{{ url_for('static', filename='jquery-1.8.3.min.js') }}"></script>
+		<script type="text/javascript">
+		$(function () {
+    $.getJSON('http://127.0.0.1:5000/MH_Z19T.json', function (MH_Z19T)
+     {
+        // Create the chart
+        var chart1 = $('#container').highcharts('StockChart', {
+            rangeSelector : {
+                selected : 1
+            },
+            title : {
+                text : 'MH_Z19 온도'
+            },
+            series : [{
+                name : 'Value',
+                data : MH_Z19T,
+                tooltip: {
+                    valueDecimals: 2
+                }
+            }]
+        });
+    });
+});
+		</script>
+	</head>
+	<body>
+		<script src="{{ url_for('static', filename='highstock.js') }}"></script>
+		<script src="{{ url_for('static', filename='highcharts-more.js') }}"></script>
+		<script src="{{ url_for('static', filename='exporting.js') }}"></script>
+		<div id="container" style="min-width: 310px; height: 400px; margin: 0 auto"></div>
+	</body>
+</html>
